requirements.txt
#pytest
pytest>=8.4.2

python-dotenv>=1.2.1

psycopg2-binary>=2.9.11

#django
django>=5.2.8
djangorestframework>=3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers>=4.9.0
django-axes>=8.0.0
django-otp>=1.6.3
argon2-cffi>=25.1.0
pip-audit>=2.9.0
drf-spectacular>=0.27.0